stable_diffusion_webapp/requirements.txt
Flask==2.3.3
torch==2.0.1
diffusers==0.21.4
transformers==4.34.0
accelerate==0.23.0
Pillow==10.0.1
gunicorn==21.2.0
requests==2.31.0
numpy==1.24.3
opencv-python==4.8.1.78
flask-cors==4.0.0